+++ packages/contracts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Rename `SpendTransaction.noteCommitment` to `newNoteCommitment` for clarity
 - Add tests for verifier contract
 - Make commitment tree and hash functions generic behind interfaces
 - Add contracts as package in yarn workspace